backend_services/shared_definitions/requirements.txt
pydantic==2.3.0
pyjwt[crypto]==2.8.0
